apply overdue formatting
this is subject to change…

Especially for lua files the LSP formatter and stylua do not agree,
therefore formatting might be jumping around a lot for no reason.

I either have to find a way to make them produce the same style, or a
way to use the LSPs formatting from the CLI, such that I can use it from
a script for an eventual hook.